added getLogsURL function to all activities
