+++ app/src/main/java/edu/rpi/shuttletracker/AboutActivity.kt
@@ -1,6 +1,8 @@
 package edu.rpi.shuttletracker
 
+import android.content.Context
 import android.content.Intent
+import android.content.SharedPreferences
 import android.content.res.Resources
 import android.net.Uri
 import androidx.appcompat.app.AppCompatActivity
@@ -8,6 +10,7 @@ import android.os.Bundle
 import android.view.MenuItem
 import android.widget.Button
 import androidx.appcompat.widget.Toolbar
+import java.net.URL
 
 class AboutActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -24,6 +27,7 @@ class AboutActivity : AppCompatActivity() {
                 Intent.ACTION_VIEW,
                 Uri.parse(res.getString(R.string.repo_url))
             )
+            Logs.writeToLogBuffer(object{}.javaClass.enclosingMethod.name, "opened browser to github link")
             startActivity(browserIntent)
         }
 
@@ -44,4 +48,14 @@ class AboutActivity : AppCompatActivity() {
         return super.onOptionsItemSelected(item)
     }
 
+    // for future logging
+    private fun getLogsURL(): URL {
+        val res : Resources = getResources()
+        val sharedPreferences: SharedPreferences =
+            getSharedPreferences("preferences", Context.MODE_PRIVATE)
+        val server_url = sharedPreferences.getString("server_base_url", res.getString(R.string.default_server_url))
+        val logsUrl =
+            URL(server_url + res.getString(R.string.logs_url))
+        return logsUrl
+    }
 }
